rename integration test package
